refactor(zipextractor): organise zipextractor
