+++ src/core/usecase/ZipExtractor/IZipExtractor.ts
